Add JWT verification to get the login user like and collect status
