+++ src/quick-view/schema/like.schema.ts
