--- src/quickView/types/post-data.type.ts
@@ -1,10 +1,0 @@
-export type Post_Data = {
-  user_name;
-  user_id: string;
-  like_count: number;
-  like_status: boolean;
-  collect_count: number;
-  collect_status: boolean;
-  photo_url: string;
-  avatar_url: string;
-};
